Applied fixes from code review
- Removed WebTorrentStreamInfo, now its StreamInfo
- Some fixes on streamer
- Removed script call on index
