+++ src/app/index.html
@@ -84,7 +84,6 @@
 	<script src="lib/models/generic_collection.js"></script>
 	<script src="lib/models/movie_collection.js"></script>
 	<script src="lib/models/stream_info.js"></script>
-	<script src="lib/models/webtorrent_stream_info.js"></script>
 	<script src="lib/models/show_collection.js"></script>
   	<script src="lib/models/anime_collection.js"></script>
 	<script src="lib/models/indie_collection.js"></script>
